movie_booking_system/movie_app/requirements.txt
django
djangorestframework
mongoengine
django-environ
bcrypt
djangorestframework-simplejwt